# refactor js code into object modules
+ encapsulate get_jsonp functions into a module. works!
use callback technology to make synchronous order calls.

+ translate `tip` tag to an specific node hex.
create a reference to `tip` tag.
`tip` tag will be translated in the enter function.
(it will *not* be translated in the first time, but acturally work fine.)

+ the list are set and shown, but the anchors are bind now.
(**idea below are NOT implemented**)
i should judge *how to bind anchor to the corrent link*.
in now stage, the anchor infomation in the view and not use suffcient.

i am thinking, not just use `javascript:void(0)` for anchor link,
use some actual function name and parameter.

+ the history push and pop are working now.
but, **DONNOT** know what should be shown on the revision anchor for every file.
maybe, a dialog? or link to the specified changeset blog.

+ the order of the list in sources can be improved. do it later.
(**TODO**)

+ fix a little bug, modify CSS for sources table list.
fix CSS in changeset file list.

+ rename util.js into hint.js.

+ finish all parts of refactor.