# fix the bug about overlay div
when popstate, judge the type of the view.

if that is *file*, do *NOT* bind anchor on list,
because when click the overlay divison, `bindList()` function
will be call again. then, two listeners bind on one single
anchor, bug appear.

now, when the type is file, just show file and bind file.
if the type is folder, bind the list.